Graph, Node, and Edge classes that implement width, depth, and shortest path algorithms.
Represents the use of graphs and search algorithms for those graphs. Graphs are made up of Nodes, and Nodes are connected using Edges. g1.txt is the input to the graph to initialize it. Edges can also include weights, which will be taken into account when searching for shortest path.